Sesiones agregadas al backend con tokens de authenticacion y rutas de gestion de clientes protegidas

+++ routes/api.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\SolicitudesController as Solicitudes;
 use App\Http\Controllers\ClienteController as Cliente;
 use App\Http\Controllers\DireccionClienteController as DireccionCliente;
 use App\Http\Controllers\SucursalController as Sucursal;
+use App\Http\Controllers\UsuarioController as Usuario;
 
 /*
 |--------------------------------------------------------------------------
@@ -22,47 +23,53 @@ use App\Http\Controllers\SucursalController as Sucursal;
 */
 
 ###########################################
+########    RUTAS DE USUARIOS   ###########
+###########################################
+Route::post('/api/login', [Usuario::class, 'login']);
+
+Route::post('/api/logout', [Usuario::class, 'logout'])->middleware(['auth:sanctum']);
+###########################################
 ########    RUTAS DE EMPRESAS   ###########
 ###########################################
-Route::get('/api/empresas',  [Empresa::class, 'getAllEmpresas']);
+Route::get('/api/empresas',  [Empresa::class, 'getAllEmpresas'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
 
 Route::get('/api/empresas/{id}',  [Empresa::class, 'getEmpresa']);
 
 Route::post('/api/empresas',  [Empresa::class, 'guardarEmpresa']);
 
-Route::put('/api/empresas/{id}',  [Empresa::class, 'actualizarEmpresa']);
+Route::put('/api/empresas/{id}',  [Empresa::class, 'actualizarEmpresa'])->middleware(['auth:sanctum', 'ability:empresa']);
 
-Route::delete('/api/empresas/{id}',  [Empresa::class, 'eliminarEmpresa']);
+Route::delete('/api/empresas/{id}',  [Empresa::class, 'eliminarEmpresa'])->middleware(['auth:sanctum', 'ability:empresa']);
 ###########################################
 ########    RUTAS DE SUCURSALES   ###########
 ###########################################
 Route::get('/api/sucursales/{id}', [Sucursal::class, 'getSucursales']);
-Route::post('/api/sucursales', [Sucursal::class, 'guardarSucursal']);
-Route::put('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'actualizarSucursal']);
-Route::delete('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'eliminarSucursal']);
+Route::post('/api/sucursales', [Sucursal::class, 'guardarSucursal'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
+Route::put('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'actualizarSucursal'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
+Route::delete('/api/sucursales/{empresa_id}/{direccion_id}', [Sucursal::class, 'eliminarSucursal'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
 ###########################################
 ########    RUTAS DE SOLICITUDES   ########
 ###########################################
-Route::get('/api/solicitudes',  [Solicitudes::class, 'getAllSolicitudes']);
+Route::get('/api/solicitudes',  [Solicitudes::class, 'getAllSolicitudes'])->middleware(['auth:sanctum', 'ability:admin']);
 
-Route::put('/api/solicitudes/{id}',  [Solicitudes::class, 'actualizarSolicitud']);
+Route::put('/api/solicitudes/{id}',  [Solicitudes::class, 'actualizarSolicitud'])->middleware(['auth:sanctum', 'ability:admin']);
 
 //se rechaza la empresa y se elimina xd
-Route::delete('/api/solicitudes/{id}',  [Solicitudes::class, 'rechazarSolicitud']);
+Route::delete('/api/solicitudes/{id}',  [Solicitudes::class, 'rechazarSolicitud'])->middleware(['auth:sanctum', 'ability:admin']);
 
 ###########################################
 ########    RUTAS DE Clientes    ##########
 ###########################################
 
-Route::get('/api/clientes',  [Cliente::class, 'getAllClientes']);
+Route::get('/api/clientes',  [Cliente::class, 'getAllClientes'])->middleware(['auth:sanctum', 'ability:empresa, admin']);
 
 Route::get('/api/clientes/{id}',  [Cliente::class, 'getCliente']);
 
-Route::post('/api/clientes',  [Cliente::class, 'guardarCliente']);
+Route::post('/api/clientes',  [Cliente::class, 'guardarCliente'])->middleware(['auth:sanctum', 'ability:empresa']);
 
-Route::put('/api/clientes/{id}',  [Cliente::class, 'actualizarCliente']);
+Route::put('/api/clientes/{id}',  [Cliente::class, 'actualizarCliente'])->middleware(['auth:sanctum', 'ability:admin,cliente,empresa']);
 
-Route::delete('/api/clientes/{id}',  [Cliente::class, 'eliminarCliente']);
+Route::delete('/api/clientes/{id}',  [Cliente::class, 'eliminarCliente'])->middleware(['auth:sanctum', 'ability:cliente,empresa']);
 
 ###########################################
 ##    RUTAS DE Direccion Clientes    ######
@@ -70,11 +77,11 @@ Route::delete('/api/clientes/{id}',  [Cliente::class, 'eliminarCliente']);
 
 Route::get('/api/clientes/{id}/direcciones',  [DireccionCliente::class, 'getClienteDirecciones']);
 
-Route::post('/api/clientes/{id}/direcciones',  [DireccionCliente::class, 'guardarClienteDireccion']);
+Route::post('/api/clientes/{id}/direcciones',  [DireccionCliente::class, 'guardarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente']);
 
-Route::put('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'actualizarClienteDireccion']);
+Route::put('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'actualizarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente']);
 
-Route::delete('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'eliminarClienteDireccion']);
+Route::delete('/api/clientes/{id}/direcciones/{id_direccion}',  [DireccionCliente::class, 'eliminarClienteDireccion'])->middleware(['auth:sanctum', 'ability:cliente']);
 
 ###########################################
 ########    RUTAS DE RECLAMOS    ##########
